Update import statements to work with python 2.4
This further updates the import statements to work with python 2.4.
Brief testing indicates that everything else seems to work correctly now
in 2.4.

I haven't fully thought out plans for which versions of python we want
to support.  Given that there are still plenty of machines running 2.4,
and that the changes needed to support it are minimal, there seems to be
no reason not to do this.